chore: refactor conformance test runners
Separate conformance runners for 0.1.x and 0.2.x test suites
