+++ conformance/conformance-baseline-0.1.yml
@@ -1,0 +1,2 @@
+server:
+  - none
