+++ conformance/src/test/java/dev/tachyonmcp/conformance/DefaultServerConformanceIT.java
@@ -4,20 +4,12 @@
 
 package dev.tachyonmcp.conformance;
 
-import dev.tachyonmcp.server.McpServer;
-import dev.tachyonmcp.server.TachyonServer;
-
 /**
  * Test against the latest stable version of MCP conformance tests
  */
 class DefaultServerConformanceIT extends AbstractServerConformanceIT {
 
     DefaultServerConformanceIT() {
-        super("0.1.16");
-    }
-
-    @Override
-    protected McpServer createServer() {
-        return TachyonServer.builder().network(n -> n.host("localhost")).build();
+        super(new DefaultConformanceServer(), "0.1.16", "conformance-baseline-0.1.yml", "default", "2025-11-25");
     }
 }
