+++ conformance/src/test/java/dev/tachyonmcp/conformance/EdgeServerConformanceIT.java
@@ -4,20 +4,12 @@
 
 package dev.tachyonmcp.conformance;
 
-import dev.tachyonmcp.server.McpServer;
-import dev.tachyonmcp.server.TachyonServer;
-
 /**
  * Test against the latest edge version of MCP conformance tests
  */
 class EdgeServerConformanceIT extends AbstractServerConformanceIT {
 
     EdgeServerConformanceIT() {
-        super("0.2.0-alpha.4");
-    }
-
-    @Override
-    protected McpServer createServer() {
-        return TachyonServer.builder().network(n -> n.host("localhost")).build();
+        super(new EdgeConformanceServer(), "0.2.0-alpha.7", "conformance-baseline-0.2.yml", "edge", "2026-07-28");
     }
 }
