+++ conformance/src/test/resources/junit-platform.properties
@@ -1,2 +1,4 @@
+junit.jupiter.execution.parallel.enabled = true
+junit.jupiter.execution.parallel.mode.default = concurrent
 junit.jupiter.execution.parallel.config.executor-service=WORKER_THREAD_POOL
 junit.jupiter.execution.timeout.default = 1 m
